requirements.txt
pyautogen==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
